refactor(authenticationLocation): moved authentication files to separate locations
