+++ src/main/java/pl/uj/passgo/configuration/security/baseAuth/BaseAuthConfiguration.java
@@ -1,4 +1,4 @@
-package pl.uj.passgo.configuration.security;
+package pl.uj.passgo.configuration.security.baseAuth;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
